♻️ Refactor chat sections.
