+++ src/pages/chat/Chat.tsx
@@ -1,21 +1,16 @@
 "use client";
 
-import { ChatBodySection } from "@/components/chat/ChatBodySection";
-import { InputSection } from "@/components/chat/InputSection";
-import { ReceivedMessage } from "@/components/chat/ReceivedMessage";
-import { SendedMessage } from "@/components/chat/SendedMessage";
-import { SideSection } from "@/components/side-section/SideSection";
+import { RoomProvider } from "@/context/room/Room.provider";
 import { useSocketContext } from "@/context/socket/Socket.context";
 import { useUserContext } from "@/context/user/User.context";
-import { useEffect, useState } from "react";
-import { RoomButton } from "./components/rooms/RoomButton";
-import { UserSection } from "@/components/side-section/UserSection";
-import { JoinRoomModal } from "./components/rooms/JoinRoomModal";
-import { RoomList } from "./components/rooms/RoomList";
-import { NewRoomModal } from "./components/rooms/NewRoomModal";
-import { roomService } from "@/services/room/roomService";
 import { MessageView } from "@/domain/interfaces/message/message";
-import { mapMessageResponseToMessageView } from "@/domain/mappers/message.mapper";
+import { ChatRoom } from "@/domain/interfaces/room/chatRoom";
+import { ChatBodySection } from "@/pages/chat/sections/current-chat/chat/ChatBodySection";
+import { InputSection } from "@/pages/chat/sections/current-chat/chat/components/InputSection";
+import { ReceivedMessage } from "@/pages/chat/sections/current-chat/chat/components/ReceivedMessage";
+import { SendedMessage } from "@/pages/chat/sections/current-chat/chat/components/SendedMessage";
+import { useEffect, useState } from "react";
+import { SideSection } from "./sections/side/chats/SideSection";
 
 
 
@@ -23,15 +18,12 @@ import { mapMessageResponseToMessageView } from "@/domain/mappers/message.mapper
 export function ChatPage() {
   const socketProvided = useSocketContext();
   const { user } = useUserContext()
-  const [selectedRoom, setSelectedRoom] = useState("")
+  const [selectedRoom, setSelectedRoom] = useState<ChatRoom>()
   const [messages, setMessages] = useState<MessageView[]>([]);
-  const [showNewRoom, setShowNewRoom] = useState<boolean>(false)
-  const [showJoinRoom, setShowJoinRoom] = useState<boolean>(false)
   useEffect(() => {
     receiveMessages();
     socketProvided && socketProvided.socket.emit("joinRoom", "BC5QVX");
   }, []);
-  console.log(messages);
   const receiveMessages = () => {
     if (!socketProvided?.socket) return;
     socketProvided.socket.on("newMessage", ({ message, sender, senderSocketId }) => {
@@ -47,6 +39,7 @@ export function ChatPage() {
       ]);
     });
   };
+
   const sendMessage = async (textToSend: string) => {
     const createMessage = await fetch("api/message", {
       method: 'POST',
@@ -68,53 +61,33 @@ export function ChatPage() {
     return true
   }
 
-  const handleSelectChat = async (roomCodeSelected:string) => {
-    setSelectedRoom(roomCodeSelected)
-    const chatRoomData = await roomService.getRoomData(roomCodeSelected)
-    if(!chatRoomData) return
-    setMessages(mapMessageResponseToMessageView(chatRoomData.messages, user))
-  }
-
-  const showCreateRoomModal = (show:boolean) => {
-    setShowNewRoom(show)
-  }
-
-  const showJoinRoomModal = (show:boolean) => {
-    setShowJoinRoom(show)
-  }
 
   return (
-    <section className="flex gap-2 w-full h-full">
-      <SideSection>
-        <div className="flex flex-1 flex-col gap-2 overflow-hidden">
-          <RoomButton showJoinRoomModal={showJoinRoomModal} showCreateRoomModal={showCreateRoomModal} />
-          <RoomList handleSelectChat={handleSelectChat}/>
-        </div>
-        <UserSection />
-      </SideSection>
-      {selectedRoom ? 
-        <div className="flex flex-col gap-2 h-full w-full">
-          <ChatBodySection>
-            {messages.length > 0 &&
-              messages.map((message, iter) =>
-                message.type === "received" && message.from !== user.username ? (
-                    <ReceivedMessage senderName={message.from} key={iter} lastMessageSent={handleLastMessage(iter)}>
-                      {message.message}
-                    </ReceivedMessage>
-                ) : (
-                    <SendedMessage key={iter} lastMessageSent={handleLastMessage(iter)} >{message.message}</SendedMessage>
+    <RoomProvider>
+      <section className="flex gap-2 w-full h-full">
+        <SideSection/>
+        {selectedRoom ? 
+          <div className="flex flex-col gap-2 h-full w-full">
+            <ChatBodySection>
+              {messages.length > 0 &&
+                messages.map((message, iter) =>
+                  message.type === "received" && message.from !== user.username ? (
+                      <ReceivedMessage senderName={message.from} key={iter} lastMessageSent={handleLastMessage(iter)}>
+                        {message.message}
+                      </ReceivedMessage>
+                  ) : (
+                      <SendedMessage key={iter} lastMessageSent={handleLastMessage(iter)} >{message.message}</SendedMessage>
 
-                )
-              )}
-          </ChatBodySection>
-          <InputSection sendMessage={sendMessage} />
-        </div>:
-        <div className="rounded-2xl flex justify-center items-center overflow-hidden p-4 gap-2 border-[1px] bg-[#121212]  border-custom-border-color w-full text-custom-text-color flex-1">
-          <span>Selecciona un chat!</span>
-        </div>
-      }
-      {showNewRoom && <NewRoomModal handleIsOpenModal={showCreateRoomModal} isOpen={showNewRoom}/>}
-      {showJoinRoom && <JoinRoomModal handleIsOpenModal={showJoinRoomModal} isOpen={showJoinRoom}/>}
-    </section>
+                  )
+                )}
+            </ChatBodySection>
+            <InputSection sendMessage={sendMessage} />
+          </div>:
+          <div className="rounded-2xl flex justify-center items-center overflow-hidden p-4 gap-2 border-[1px] bg-[#121212]  border-custom-border-color w-full text-custom-text-color flex-1">
+            <span>Selecciona un chat!</span>
+          </div>
+        }
+      </section>
+    </RoomProvider>
   );
 }
